+++ NEWS.md
@@ -1,3 +1,13 @@
+# semhelpinghands 0.1.9.1
+
+## New Features
+
+- Added `ptable_to_syntax]()` for
+  converting a `lavaan` parameter
+  table to a `lavaan` model syntax.
+  (0.1.9.1)
+
+
 # semhelpinghands 0.1.9
 
 ## New Features